Add a prompt to delete-job about cancel-job

+++ lib/flight_job/commands/delete_job.rb
@@ -25,14 +25,35 @@
 # https://github.com/openflighthpc/flight-job
 #==============================================================================
 
+require 'tty-prompt'
+
 module FlightJob
   module Commands
     class DeleteJob < Command
       def run
         job = load_job(args.first)
+
+        if $stdout.tty? && !job.terminal?
+          $stderr.puts pastel.red <<~WARN.chomp
+            Job '#{job.id}' is currently #{job.state}!
+            You should first cancel the job with:
+          WARN
+          $stderr.puts pastel.yellow "#{CLI.program(:name)} cancel-job #{args.first}"
+          $stderr.puts
+          if prompt.no? pastel.bold("Do you wish to delete the job anyway?")
+            raise InputError, "The job has not been deleted"
+          end
+        end
+
         FlightJob.logger.warn "Permanently removing job: #{job.id}"
         FileUtils.rm_rf File.dirname(job.metadata_path)
         $stderr.puts "Removed job: #{job.id}"
+      end
+
+      private
+
+      def prompt
+        @prompt ||= TTY::Prompt.new(help_color: :yellow)
       end
     end
   end
